+++ Servers/.settings/org.eclipse.wst.server.core.prefs
@@ -1,3 +1,3 @@
-#Wed Dec 04 13:54:44 KST 2024
-org.eclipse.wst.server.core.isServerProject=true
+#Mon Dec 09 11:22:07 KST 2024
 eclipse.preferences.version=1
+org.eclipse.wst.server.core.isServerProject=true
