--- demo/src/main/java/com/example/demo/controller/HomeController.java
@@ -1,17 +1,0 @@
-package com.example.demo.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class HomeController {
-    @GetMapping("/")
-    public String index() {
-        return "index";
-    }
-    
-    @GetMapping("/example")
-    public String example() {
-    	return "example/example";
-    }
-}
