--- demo/src/main/resources/application.properties
@@ -1,20 +1,0 @@
-spring.application.name=demo
-
-#port
-server.port=8080
-
-#encoding
-server.servlet.encoding.charset=UTF-8
-server.servlet.encoding.enabled=true
-server.servlet.encoding.force=true
-
-#restart
-spring.devtools.livereload.enabled=true
-spring.devtools.restart.enabled=true
-
-#static resources
-spring.web.resources.static-locations=classpath:/static/
-spring.mvc.static-path-pattern=/css/**,/js/**,/images/**
-
-#db nolink
-spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
