--- demo/src/main/resources/templates/fragments/default.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
-<head>
-    <meta charset="UTF-8">
-    <title>default layout</title>
-	<link rel="stylesheet" href="/css/bootstrap.min.css" th:href="@{/css/bootstrap.min.css}" />
-	<link rel="stylesheet" href="/css/style.css" th:href="@{/css/style.css}" />
-    <script th:src="@{/js/bootstrap.min.js}"></script>
-</head>
-<body>
-    <header th:replace="fragments/header :: header"></header>
-    
-    <div layout:fragment="content">
-        <!-- 각 페이지의 컨텐츠가 여기에 들어갑니다 -->
-    </div>
-    
-    <footer th:replace="fragments/footer :: footer"></footer>
-</body>
-</html>
