--- demo/src/main/resources/templates/fragments/footer.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<body>
-    <footer th:fragment="footer">
-        <div class="d-flex align-items-center justify-content-center column-gap-5">
-			<div class="d-block">
-				<img th:src="@{/static/images/footerlogo.jpg}" alt="Footer Logo" width="140" height="100">
-			</div>
-			<div class="w-50">
-				<p><strong>(주)빌드이지</strong></p>
-				<p>주소: 서울특별시 강남구 언주로 508 14층 (역삼동, 서울상록빌딩) 대표: 김기현, 나경수, s이미은, 정병국, 황혜령</p>
-				<p>사업자번호: 104-81-53114 | 통신판매업: 제2016-서울강남-00151호 | 부가통신사업: 제112233호 <a href="#">[사업자정보확인]</a></p>
-				<p>(주)빌드이지는 상품판매와 직접적인 관련이 없으며, 모든 상거래의 책임은 구매자와 판매자에게 있습니다. 이에 대해 (주)빌드이지는 일체의 책임을 지지 않습니다. 본사에 등록된 모든 광고와 저작권 및 법적책임은 자료제공사 (또는 글쓴이)에게 있으므로 본사는 광고에 대한 책임을 지지 않습니다.</p>
-				<p class="text-muted small">Copyright © buildeasy Co., Ltd. All Rights Reserved.</p>
-			</div>
-		</div>
-    </footer>
-</body>
-</html>
