--- demo/src/main/resources/templates/fragments/header.html
@@ -1,103 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head th:fragment="headerFragment">
-    <meta charset="UTF-8">
-    <title>EasyBuild2</title>
-	<link rel="stylesheet" href="/css/bootstrap.min.css" th:href="@{/css/bootstrap.min.css}" />
-	<link rel="stylesheet" href="/css/style.css" th:href="@{/css/style.css}" />
-	<script th:src="@{/js/bootstrap.min.js}"></script>
-</head>
-<body>
-    <header th:fragment="header">
-  		<div class="d-flex justify-content-end pe-5 bg-dark">
-			<div class="p-2">
-				<a class="nav-link text-light" aria-current="page" href="#">마이페이지</a>
-			</div>
-			<div class="p-2">
-				<a class="nav-link text-light" href="#">로그인</a>
-			</div>
-		</div>
-		<div class="position-relative d-flex justify-content-center py-5">
-			<!-- 로고 -->
-			<div class="position-absolute logo-box">
-	          <a th:href="@{/}" class="d-block w-100">
-	            <img th:src="@{/images/logo.png}" alt="Logo" width="140" height="100">
-	          </a>
-			</div>
-	        <!-- 검색바 -->
-	        <form class="w-50" role="search" th:action="@{/search}" method="get">
-	          <input type="search" class="form-control" placeholder="검색어를 입력하세요." aria-label="Search">
-	        </form>			
-		</div>
-		<!-- 네비게이션바 -->
-		<nav class="d-flex justify-content-between position-relative text-center main-nav">
-			<div class="nav-col">
-				<a href="javascript: void(0)" class="nav-main-depth text-black">pc주요부품</a>
-				<div class="position-absolute nav-sub-depth">
-					<!-- 서브 뎁스는 3개씩 묶어서 추가 -->
-					<div class="d-flex flex-column row-gap-5 nav-sub-depth-col">
-						<a href="javascript: void(0)" class="text-black">pc주요부품1</a>
-						<a href="javascript: void(0)" class="text-black">pc주요부품2</a>
-						<a href="javascript: void(0)" class="text-black">pc주요부품3</a>
-					</div>
-					<div class="d-flex flex-column row-gap-5 nav-sub-depth-col">
-						<a href="javascript: void(0)" class="text-black">pc주요부품4</a>
-						<a href="javascript: void(0)" class="text-black">pc주요부품5</a>
-						<a href="javascript: void(0)" class="text-black">pc주요부품6</a>
-					</div>
-				</div>
-			</div>
-			<div class="nav-col">
-				<a href="javascript: void(0)" class="nav-main-depth text-black">모니터</a>
-				<div class="position-absolute nav-sub-depth">
-					<div class="d-flex flex-column row-gap-5 nav-sub-depth-col">
-						<a href="javascript: void(0)" class="text-black">모니터1</a>
-						<a href="javascript: void(0)" class="text-black">모니터2</a>
-						<a href="javascript: void(0)" class="text-black">모니터3</a>
-					</div>
-				</div>
-			</div>
-			<div class="nav-col">
-				<a href="javascript: void(0)" class="nav-main-depth text-black">키보드</a>
-				<div class="position-absolute nav-sub-depth">
-					<div class="d-flex flex-column row-gap-5 nav-sub-depth-col">
-						<a href="javascript: void(0)" class="text-black">키보드1</a>
-						<a href="javascript: void(0)" class="text-black">키보드2</a>
-						<a href="javascript: void(0)" class="text-black">키보드3</a>
-					</div>
-				</div>
-			</div>
-			<div class="nav-col">
-				<a href="javascript: void(0)" class="nav-main-depth text-black">마우스</a>
-				<div class="position-absolute nav-sub-depth">
-					<div class="d-flex flex-column row-gap-5 nav-sub-depth-col">
-						<a href="javascript: void(0)" class="text-black">마우스1</a>
-						<a href="javascript: void(0)" class="text-black">마우스2</a>
-						<a href="javascript: void(0)" class="text-black">마우스3</a>
-					</div>
-				</div>
-			</div>
-			<div class="nav-col">
-				<a href="javascript: void(0)" class="nav-main-depth text-black">사운드</a>
-				<div class="position-absolute nav-sub-depth">
-					<div class="d-flex flex-column row-gap-5 nav-sub-depth-col">
-						<a href="javascript: void(0)" class="text-black">사운드1</a>
-						<a href="javascript: void(0)" class="text-black">사운드2</a>
-						<a href="javascript: void(0)" class="text-black">사운드3</a>
-					</div>
-				</div>
-			</div>
-			<div class="nav-col">
-				<a href="javascript: void(0)" class="nav-main-depth text-black">게시판</a>
-				<div class="position-absolute nav-sub-depth">
-					<div class="d-flex flex-column row-gap-5 nav-sub-depth-col">					
-						<a href="javascript: void(0)" class="text-black">게시판1</a>
-						<a href="javascript: void(0)" class="text-black">게시판2</a>
-						<a href="javascript: void(0)" class="text-black">게시판3</a>
-					</div>
-				</div>
-			</div>
-		</nav>
-    </header>
-</body>
-</html>
