+++ pcShoppingMall/.gitattributes
