+++ pcShoppingMall/src/main/java/com/project/easyBuild/PcShoppingMallApplication.java
@@ -1,0 +1,13 @@
+package com.project.easyBuild;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PcShoppingMallApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PcShoppingMallApplication.class, args);
+	}
+
+}
