+++ pcShoppingMall/src/main/java/com/project/easyBuild/config/WebConfig.java
@@ -1,0 +1,18 @@
+package com.project.easyBuild.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@EnableWebMvc
+@ComponentScan("controllers")
+public class WebConfig implements WebMvcConfigurer {
+	@Override
+	public void addResourceHandlers(ResourceHandlerRegistry registry) {
+	    registry.addResourceHandler("/**")
+	            .addResourceLocations("classpath:/static/");
+	}
+}
