+++ pcShoppingMall/src/main/java/com/project/easyBuild/controller/HomeController.java
@@ -1,0 +1,17 @@
+package com.project.easyBuild.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+    @GetMapping("/")
+    public String index() {
+        return "index";
+    }
+    
+    @GetMapping("/example")
+    public String example() {
+    	return "example/example";
+    }
+}
