+++ pcShoppingMall/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,0 +1,5 @@
+{"properties": [{
+  "name": "spring.mvc.view.viewClass",
+  "type": "java.lang.String",
+  "description": "A description for 'spring.mvc.view.viewClass'"
+}]}
