+++ pcShoppingMall/src/main/resources/application.properties
@@ -1,0 +1,19 @@
+spring.application.name=demo
+
+#port
+server.port=8080
+
+#encoding
+server.servlet.encoding.charset=UTF-8
+server.servlet.encoding.enabled=true
+server.servlet.encoding.force=true
+
+#restart
+spring.devtools.livereload.enabled=true
+spring.devtools.restart.enabled=true
+
+#static resources
+spring.mvc.static-path-pattern=/**
+
+#db nolink
+spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
