+++ pcShoppingMall/src/main/resources/static/css/style.css
