+++ pcShoppingMall/src/test/java/com/project/easyBuild/PcShoppingMallApplicationTests.java
@@ -1,0 +1,13 @@
+package com.project.easyBuild;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PcShoppingMallApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
